wait for ready status in unit tests

+++ tests.py
@@ -5,6 +5,16 @@ import os
 
 from pyomniar import *
 
+class OmniarTestError(Exception):
+    """Omniar test exception"""
+
+    def __init__(self, reason):
+        self.reason = unicode(reason)
+
+    def __str__(self):
+        return self.reason
+
+
 """Configurations"""
 # Must supply omniar account credentials for tests
 api_key = ''
@@ -25,7 +35,22 @@ class PyomniarAPITests(unittest.TestCase):
         
         self.updated_name = "updated_name_atm"
         self.updated_content = "updated_name_content"
-        
+    
+    def _wait_for_ready(self, scan_id):
+        print('waiting for scan to process')
+        loopout = 60 #(10 mins)
+        count = 0
+        while count < loopout:
+            single = self.api.get_scan(scan_id=scan_id)
+            if single['status'] == 'scan_building':
+                sleep(10)
+            else:
+                break
+            count += 1
+
+        if single['status'] != 'scan_active':
+            raise OmniarTestError('Scan did not process')
+    
     def test_account_info(self):
         '''Test call to get account info'''
         ret = self.api.get_account()
@@ -37,7 +62,7 @@ class PyomniarAPITests(unittest.TestCase):
         for r in ret:
             self.api.delete_scan(r['scanUUID'])
         print(ret)
-    
+        
     def test_post_get_update(self):
         ret = self.api.post_scans(os.getcwd()+'/test_data/atm_scan.zip', 
             name='tests_atm', 
@@ -45,9 +70,6 @@ class PyomniarAPITests(unittest.TestCase):
         ret_id = ret.get('scanUUID')
         if ret_id:
             self.scan_ids.append(ret_id)
-        print(ret_id)
-        print(ret)
-        sleep(5)
         
         # get single scan by id
         print("get single scan")
@@ -55,13 +77,16 @@ class PyomniarAPITests(unittest.TestCase):
         single_id = ret.get('scanUUID')
         self.assertEqual(ret_id, single_id)
         
+        self._wait_for_ready(ret_id)
+        
         # update scan with new data
         print("update scan")
         update = self.api.update_scan(ret_id,
             name = self.updated_name,
             content = self.updated_content
         )
-        sleep(5)
+        
+        sleep(10)
         
         print("get updated scan")
         # get updated scan by id
@@ -73,44 +98,36 @@ class PyomniarAPITests(unittest.TestCase):
 
     # def test_post_match(self):
     #         ret = self.api.post_scans(os.getcwd()+'/test_data/atm_scan.zip', 
-    #             name='tests_city_building', 
-    #             content='city building (this is a test upload for matching)')
+    #             name='tests_atm_match', 
+    #             content='atm scans (this is a test upload for matching)')
     #         ret_id = ret.get('scanUUID')
+    #         print ret_id
     #         if ret_id:
     #             self.scan_ids.append(ret_id)
-    #         print(ret)
-    #         print(ret_id)
-    #         sleep(5)
-    #         
+    #     
     #         # get single scan by id
     #         print("waiting for scan to build (this may take a while)")
     #         # wait for scan to build
-    #         loopout = 12
-    #         count = 0
-    #         while count < loopout:
-    #             single = self.api.get_scan(scan_id=ret_id)
-    #             if single['status'] == 'scan_building':
-    #                 sleep(5)
-    #             else:
-    #                 break
-    #             count += 1
-    #         
-    #         if single['status'] == 'scan_building':
-    #             self.fail("scan did not build in time")
-    # 
+    #         self._wait_for_ready(ret_id)
+    #     
     #         # do match
     #         match = self.api.match(os.getcwd()+'/test_data/atm_scan_match.jpg')
     #         # match should be a list
+    #         print("MATCH RETURNED")
     #         print(match)
     #         self.assertTrue(match)
-        
-        match_id = match[0]['scanUUID']
-        self.assertEqual(ret_id, match_id)
-
+    #         
+    #         match_id = match[0]['scanUUID']
+    #         self.assertEqual(ret_id, match_id)
+    
     def tearDown(self):
         for scan_id in self.scan_ids:
-            ret = self.api.delete_scan(scan_id=scan_id)
-            print(ret)
+            scan = self.api.get_scan(scan_id=scan_id)
+            if scan['status'] == 'scan_active':
+                ret = self.api.delete_scan(scan_id=scan_id)
+                print('scan %s deleted' % (scan_id))
+            else:
+                print("scan %s skipped, not ready" % (scan_id))
 
 if __name__ == '__main__':
     unittest.main()
